Unified logic to 1 specification that accepts an array of kontexte
